requirements-dev.txt
# Development dependencies
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.5.1
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pre-commit==3.3.3

# Additional development tools
ipython==8.14.0
jupyter==1.0.0
python-dotenv==1.1.1

# Testing utilities
factory-boy==3.3.0
freezegun==1.2.2
